Remove valid_params_request? method
This method is causing so much trouble for us when logging in via the
login modal. If we make a POST request after logging in Rails tells us
our CSRF token is not authentic.

I’m removing this method to play around with Devise’s code and find a
potential solution.